refactor: security features
