+++ core/src/main/kotlin/me/rhunk/snapenhance/core/SnapEnhance.kt
@@ -223,35 +223,7 @@ class SnapEnhance {
             }
         }
 
-        appContext.config.experimental.nativeHooks.customSharedLibrary.get().takeIf { it.isNotEmpty() }?.let {
-            runCatching {
-                appContext.native.loadSharedLibrary(
-                    appContext.fileHandlerManager.getFileHandle(FileHandleScope.USER_IMPORT.key, it).toWrapper().readBytes()
-                )
-                appContext.log.verbose("loaded custom shared library")
-            }.onFailure {
-                appContext.log.error("Failed to load custom shared library", it)
-            }
-        }
-
-        if (appContext.bridgeClient.getDebugProp("disable_sif", "false") != "true") {
-            runCatching {
-                appContext.native.loadSharedLibrary(
-                    appContext.fileHandlerManager.getFileHandle(FileHandleScope.INTERNAL.key, InternalFileHandleType.SIF.key)
-                        .toWrapper()
-                        .readBytes()
-                        .takeIf {
-                            it.isNotEmpty()
-                        } ?: throw IllegalStateException("buffer is empty")
-                )
-                appContext.log.verbose("loaded sif")
-            }.onFailure {
-                safeMode = true
-                appContext.log.error("Failed to load sif", it)
-            }
-        } else {
-            appContext.log.warn("sif is disabled")
-        }
+        val safeMode = SecurityFeatures(appContext).init()
 
         Runtime::class.java.findRestrictedMethod {
             it.name == "loadLibrary0" && it.parameterTypes.contentEquals(
