+++ core/src/main/kotlin/me/rhunk/snapenhance/core/features/FeatureManager.kt
@@ -28,6 +28,10 @@ class FeatureManager(
     private val features = mutableMapOf<KClass<out Feature>, Feature>()
     private val onActivityCreateListeners = mutableListOf<(Activity) -> Unit>()
 
+    fun addActivityCreateListener(block: (Activity) -> Unit) {
+        onActivityCreateListeners.add(block)
+    }
+
     private fun register(vararg featureList: Feature) {
         if (context.bridgeClient.getDebugProp("disable_feature_loading") == "true") {
             context.log.warn("Feature loading is disabled")
@@ -61,7 +65,6 @@ class FeatureManager(
     fun init() {
         register(
             Debug(),
-            SecurityFeatures(),
             EndToEndEncryption(),
             ScopeSync(),
             PreventMessageListAutoScroll(),
